pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pom_project"
version = "0.1.0"
description = "Playwright Page Object Model example project"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Your Name", email = "you@example.com" }]
dependencies = []

[project.optional-dependencies]
dev = [
  "playwright>=1.54,<2.0",
  "pytest>=8.3,<9.0",
  "pytest-playwright>=0.7,<1.0",
  "pytest-html>=4.1,<5.0",
  "allure-pytest>=2.13,<3.0",
  "python-dotenv>=1.0,<2.0",
  "ruff>=0.8,<1.0",
  "mypy>=1.13,<2.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.11"
files = ["src"]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
